Don't store log attachments as BLOB in db
Loading big BLOBs will cause OutOfMemoryError, so instead of storing
attachments as BLOBs in the database, save them as files in the app
storage, and use the filename to retrieve them.

+++ app/src/main/java/de/spiritcroc/ownlog/Util.java
@@ -29,7 +29,7 @@ public final class Util {
     // No instantiation wanted
     private Util() {}
 
-    public static String formatFileSize(Context context, int size) {
+    public static String formatFileSize(Context context, long size) {
         if (size > FILESIZE_GIB) {
             return context.getString(R.string.filesize_gib, size/FILESIZE_GIB);
         } else if (size > FILESIZE_MIB) {
